Feat[Avatar]: moved Avatar component to UI directory
